+++ packages/js/settings-editor/src/index.tsx
@@ -19,8 +19,9 @@ import { store as editSiteStore } from '@wordpress/edit-site/build-module/store'
  */
 import { isGutenbergVersionAtLeast } from './utils';
 import { Layout } from './layout';
-import { useActiveRoute } from './route';
-import { useRegisterSiteEditorRoutes } from './components/routes';
+import { useActiveRoute } from './routes/route';
+import useRegisterSiteEditorRoutes from './hooks/use-register-site-editor-routes';
+import settingsEditorRoutes from './routes';
 
 const { RouterProvider } = unlock( routerPrivateApis );
 
@@ -32,7 +33,7 @@ const SettingsLayout = () => {
 
 export const SettingsEditor = () => {
 	const isRequiredGutenbergVersion = isGutenbergVersionAtLeast( 19.0 );
-	useRegisterSiteEditorRoutes();
+	useRegisterSiteEditorRoutes( settingsEditorRoutes );
 
 	const routes = useSelect( ( select ) => {
 		return unlock( select( editSiteStore ) ).getRoutes();
